KOMODO-2: Remove test code from init. Fix Playground.
Removes testing parameters from main and demonstrates primitive setup and drawing in Playground.

To make sure the engine is most flexible, engine features will be
developed in the user application using the functions available from the
engine API.

The solution is then ported into the engine as a way of demonstrating
that all engine features are implemented using functions and libraries
available to any developer using the Komodo game engine as a package.